requirements.txt
pytorch
matplotlib
nunmpy
pandas
scipy
scikit-learn
seaborn
